pyproject.toml
[project]
name = "autoshorts"
version = "0.1.0"
description = "Auto-generate short-form videos from Reddit stories"
authors = [{ name = "Tucker Trost"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["praw","google-generativeai"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
